complete r13404 - use double for timecode fps.
git-svn-id: svn://localhost/ardour2/branches/3.0@13471 d708f5d6-7413-0410-9779-e7cbd77b26cf

+++ libs/ardour/ardour/session.h
@@ -361,7 +361,7 @@ class Session : public PBD::StatefulDestructible, public PBD::ScopedConnectionLi
 		return mtc_timecode_bits;   /* encoding of SMTPE type for MTC */
 	}
 
-	float timecode_frames_per_second() const;
+	double timecode_frames_per_second() const;
 	bool timecode_drop_frames() const;
 
 	/* Locations */
